move time conversion helpers to @meru/shared/time
